fix: Remote links now work properly
Remote links with nested clickable elements like a button or an SVG
would not properly submit. They would use the `target` of the event and
not the `currentTarget`. This is now fixed and has tests around it.

fix: use event.currentTarget instead of target

fix: remote-link-submissions now work for nested elements